requirements.txt
django
djangorestframework
psycopg2-binary
django-filter
python-dotenv
gunicorn
requests
flake8
